@@ -33,77 +33,130 @@
     "test-unit-coverage": "./scripts/test-unit-coverage.sh",
     "test-audit": "node ./node_modules/.bin/audit-ci --config .audit-ci.json --pass-enoaudit"
   },
-  "devDependencies": {
-    "@babel/preset-env": "^7.24.8",
-    "@graphql-tools/merge": "^9.0.4",
-    "@graphql-tools/schema": "^10.0.4",
-    "@graphql-tools/utils": "^10.3.2",
-    "@testing-library/dom": "^10.3.2",
-    "@testing-library/jest-dom": "^6.4.6",
-    "@testing-library/react": "^16.0.0",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/chai": "^4.3.16",
-    "@types/chai-as-promised": "^7.1.8",
-    "@types/deep-freeze-strict": "^1.1.2",
-    "@types/jest": "^29.5.12",
-    "@types/lodash": "^4.17.6",
-    "@types/lodash.findindex": "^4.6.9",
-    "@types/lodash.isequal": "^4.5.8",
-    "@types/lodash.set": "^4.3.9",
-    "@types/node": "^20.14.10",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@types/relay-runtime": "^17.0.1",
-    "@typescript-eslint/eslint-plugin": "^7.16.0",
-    "@typescript-eslint/parser": "^7.16.0",
-    "audit-ci": "^7.1.0",
-    "babel-jest": "^29.7.0",
-    "blob-polyfill": "^9.0.20240710",
-    "chai": "^4.4.1",
-    "chai-as-promised": "^8.0.0",
-    "chai-jest-diff": "^1.0.2",
-    "eslint": "^8.56.0",
-    "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^28.6.0",
-    "expect-type": "^0.19.0",
-    "fetch-mock": "^10.0.8",
-    "graphql": "^16.9.0",
+  "dependencies": {
+    "@apollo/client": "^3.10.8",
+    "@graphql-typed-document-node/core": "^3.1.1",
+    "@wry/caches": "^1.0.0",
+    "@wry/equality": "^0.5.6",
+    "@wry/trie": "^0.5.0",
     "graphql-tag": "^2.12.6",
-    "graphql-ws": "^5.16.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-junit": "^16.0.0",
-    "jsdom": "^24.1.0",
-    "lodash": "^4.17.21",
-    "react": "^18.3.1",
-    "react-error-boundary": "^4.0.13",
+    "hoist-non-react-statics": "^3.3.2",
+    "deep-freeze-strict": "^1.1.1",
+    "make-error": "^1.3.6",
+    "optimism": "^0.18.0",
+    "prop-types": "^15.7.2",
     "rehackt": "^0.1.0",
-    "relay-runtime": "^17.0.0",
-    "rimraf": "^6.0.1",
-    "rxjs": "^7.8.1",
+    "response-iterator": "^0.2.6",
     "symbol-observable": "^4.0.0",
-    "ts-jest": "^29.2.2",
-    "typescript": "~5.5",
-    "use-sync-external-store": "^1.2.2",
-    "wait-for-observables": "^1.0.3",
-    "web-streams-polyfill": "^4.0.0",
+    "ts-invariant": "^0.10.3",
+    "tslib": "^2.3.0",
     "zen-observable-ts": "^1.2.5"
   },
-  "dependencies": {
-    "@apollo/client": "^3.8.4",
-    "@graphql-typed-document-node/core": "^3.2.0",
-    "@wry/equality": "^0.5.7",
-    "deep-freeze-strict": "^1.1.1",
-    "lodash.findindex": "^4.6.0",
-    "lodash.isequal": "^4.5.0",
-    "lodash.set": "^4.3.2",
-    "make-error": "^1.3.6",
-    "subscriptions-transport-ws": "^0.9.17",
-    "ts-invariant": "^0.10.3",
-    "tslib": "^2.6.3"
+  "devDependencies": {
+    "@arethetypeswrong/cli": "0.15.3",
+    "@babel/parser": "7.24.7",
+    "@babel/preset-env": "7.24.8",
+    "@changesets/changelog-github": "0.5.0",
+    "@changesets/cli": "2.27.7",
+    "@graphql-tools/merge": "9.0.4",
+    "@graphql-tools/schema": "10.0.4",
+    "@graphql-tools/utils": "10.3.1",
+    "@microsoft/api-extractor": "7.47.0",
+    "@rollup/plugin-node-resolve": "11.2.1",
+    "@size-limit/esbuild-why": "11.1.4",
+    "@size-limit/preset-small-lib": "11.1.4",
+    "@testing-library/dom": "10.3.2",
+    "@testing-library/jest-dom": "6.4.6",
+    "@testing-library/react": "15.0.7",
+    "@testing-library/react-12": "npm:@testing-library/react@^12",
+    "@testing-library/user-event": "14.5.2",
+    "@tsconfig/node20": "20.1.4",
+    "@types/bytes": "3.1.4",
+    "@types/chai": "4.3.16",
+    "@types/chai-as-promised": "7.1.8",
+    "@types/deep-freeze-strict": "1.1.2",
+    "@types/fetch-mock": "7.3.8",
+    "@types/glob": "8.1.0",
+    "@types/hoist-non-react-statics": "3.3.5",
+    "@types/jest": "29.5.12",
+    "@types/lodash": "4.17.6",
+    "@types/node": "20.14.10",
+    "@types/node-fetch": "2.6.11",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "@types/relay-runtime": "14.1.24",
+    "@types/use-sync-external-store": "0.0.6",
+    "@typescript-eslint/eslint-plugin": "7.15.0",
+    "@typescript-eslint/parser": "7.15.0",
+    "@typescript-eslint/rule-tester": "7.15.0",
+    "@typescript-eslint/types": "7.15.0",
+    "@typescript-eslint/utils": "7.15.0",
+    "acorn": "8.12.1",
+    "ajv": "8.16.0",
+    "audit-ci": "7.1.0",
+    "babel-jest": "29.7.0",
+    "blob-polyfill": "7.0.20220408",
+    "bytes": "3.1.2",
+    "chai": "4.4.1",
+    "chai-as-promised": "8.0.0",
+    "chai-jest-diff": "1.0.2",
+    "cross-fetch": "4.0.0",
+    "eslint": "8.57.0",
+    "eslint-import-resolver-typescript": "3.6.1",
+    "eslint-plugin-eslint-comments": "3.2.0",
+    "eslint-plugin-import": "npm:@phryneas/eslint-plugin-import@2.27.5-pr.2813.2817.199971c",
+    "eslint-plugin-jest": "28.6.0",
+    "eslint-plugin-local-rules": "2.0.1",
+    "eslint-plugin-react-compiler": "0.0.0-experimental-c8b3f72-20240517",
+    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-testing-library": "6.2.2",
+    "expect-type": "0.19.0",
+    "fetch-mock": "9.11.0",
+    "glob": "8.1.0",
+    "graphql": "16.9.0",
+    "graphql-tag": "2.12.6",
+    "graphql-ws": "5.16.0",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
+    "jest-junit": "16.0.0",
+    "jsdom": "24.1.0",
+    "lodash": "4.17.21",
+    "patch-package": "8.0.0",
+    "prettier": "3.1.1",
+    "react": "18.3.1",
+    "react-17": "npm:react@^17",
+    "react-19": "npm:react@19.0.0-rc-fb9a90fa48-20240614",
+    "react-dom": "18.3.1",
+    "react-dom-17": "npm:react-dom@^17",
+    "react-dom-19": "npm:react-dom@19.0.0-rc-fb9a90fa48-20240614",
+    "react-error-boundary": "4.0.13",
+    "recast": "0.23.9",
+    "resolve": "1.22.8",
+    "rehackt": "0.1.0",
+    "relay-runtime": "17.0.0",
+    "rimraf": "5.0.8",
+    "rollup": "2.79.1",
+    "rollup-plugin-cleanup": "3.2.1",
+    "rollup-plugin-terser": "7.0.2",
+    "rxjs": "7.8.1",
+    "size-limit": "11.1.4",
+    "subscriptions-transport-ws": "0.11.0",
+    "symbol-observable": "4.0.0",
+    "terser": "5.31.1",
+    "ts-api-utils": "1.3.0",
+    "ts-jest": "29.1.5",
+    "ts-jest-resolver": "2.0.1",
+    "ts-morph": "22.0.0",
+    "ts-node": "10.9.2",
+    "typedoc": "0.25.0",
+    "typescript": "5.5.3",
+    "use-sync-external-store": "1.2.2",
+    "wait-for-observables": "1.0.3",
+    "web-streams-polyfill": "4.0.0",
+    "whatwg-fetch": "3.6.20",
+    "zen-observable-ts": "1.2.5"
   },
   "peerDependencies": {
-    "@apollo/client": "^3.8.4"
+    "@apollo/client": "^3.10.8"
   }
 }
